:recycle: refactor: Rename usePrograms to useProgramsWithGates
